client: rename WithServiceToken => WithAuthToken (#1887)
